backend/requirements.txt
fastapi>=0.105.0
uvicorn[standard]>=0.23.2
python-multipart>=0.0.6
requests>=2.31.0
pydantic>=2.9.0
dlib>=19.24.0
numpy>=1.26.0
Pillow>=11.3.0
face_recognition>=1.3.0
